add validations for rolePermissionController, also created first test case for auth routes

+++ src/controllers/rolePermissionController.ts
@@ -2,11 +2,20 @@ import { NextFunction, Request, Response } from 'express';
 import Role from '../database/models/Role';
 import RolePermission from '../database/models/RolePermission';
 import { AppError } from '../middlewares/errorHandler';
+import { IRequestWithUserAndChannel } from '../types';
+import { checkPermissionLevel, findMissingPermissions, isPermissionAssignable } from '../services/permissionService';
+import { findMissingRoles } from '../services/roleService';
+import User from '../database/models/User';
 
 // Fetch all permissions assigned to a role
 const getRolePermissions = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const customReq = req as IRequestWithUserAndChannel;
         const role = await Role.findByPk(req.params.role_id);
+
+        if(!customReq.isGlobalRole && role?.channel_id != customReq.channel?.id){
+            throw new AppError("Unauthorized to view this role's permissions.", 403);
+        };
 
         if (!role) {
             throw new AppError('Role not found', 404);
@@ -26,10 +35,30 @@ const getRolePermissions = async (req: Request, res: Response, next: NextFunctio
 // Assign one or more permissions to a role
 const addRolePermissions = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const customReq = req as IRequestWithUserAndChannel;
         const role = await Role.findByPk(req.params.role_id);
+        const authorizedUser = (await User.findByPk(customReq.user.id))!;
+        const authorizeUserRoleLevel = (await checkPermissionLevel(
+            ['remove:role_permission', 'admin:role_permission'], 
+            authorizedUser
+        ))!;
 
         if (!role) {
             throw new AppError('Role not found', 404);
+        }
+
+        if(role.level <= authorizeUserRoleLevel){
+            throw new AppError("You cannot replace permissions from a role with the same level as your permission's role.", 403);
+        }
+        
+        if(!customReq.isGlobalRole && role?.channel_id != customReq.channel?.id){
+            throw new AppError("Unauthorized to add permissions to this role.", 403);
+        }
+
+        const missingPermissions = await findMissingPermissions(customReq.body.role_ids);
+        
+        if (missingPermissions.length > 0) {
+            throw new AppError(`Permission IDs ${missingPermissions} do not exist.`, 403);
         }
 
         await role.addPermissions(req.body.permission_ids);
@@ -48,10 +77,30 @@ const addRolePermissions = async (req: Request, res: Response, next: NextFunctio
 // Replace all permissions of a role
 const replaceRolePermissions = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const customReq = req as IRequestWithUserAndChannel;
         const role = await Role.findByPk(req.params.role_id);
+        const authorizedUser = (await User.findByPk(customReq.user.id))!;
+        const authorizeUserRoleLevel = (await checkPermissionLevel(
+            ['remove:role_permission', 'admin:role_permission'], 
+            authorizedUser
+        ))!;
 
         if (!role) {
             throw new AppError('Role not found', 404);
+        }
+
+        if(role.level <= authorizeUserRoleLevel){
+            throw new AppError("You cannot replace permissions from a role with the same level as your permission's role.", 403);
+        }
+
+        if(!customReq.isGlobalRole && role?.channel_id != customReq.channel?.id){
+            throw new AppError("Unauthorized to replace permissions to this role.", 403);
+        }
+
+        const missingPermissions = await findMissingPermissions(customReq.body.role_ids);
+        
+        if (missingPermissions.length > 0) {
+            throw new AppError(`Permission IDs ${missingPermissions} do not exist.`, 403);
         }
 
         await role.setPermissions(req.body.permission_ids);
@@ -70,6 +119,22 @@ const replaceRolePermissions = async (req: Request, res: Response, next: NextFun
 // Remove a specific permission from a role
 const destroyRolePermission = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const customReq = req as IRequestWithUserAndChannel;
+        const role = await Role.findByPk(customReq.params.role_id);
+        const authorizedUser = (await User.findByPk(customReq.user.id))!;
+        const authorizeUserRoleLevel = (await checkPermissionLevel(
+            ['remove:role_permission', 'admin:role_permission'], 
+            authorizedUser
+        ))!;
+
+        if (!role) {
+            throw new AppError('Role not found', 404);
+        }
+
+        if(role.level <= authorizeUserRoleLevel){
+            throw new AppError("You cannot delete a permission from a role with the same level as your permission's role.", 403);
+        }
+
         await RolePermission.destroy({
             where: {
                 permission_id: req.params.permission_id,
